+++ BetterJoy/Forms/MainForm.cs
@@ -1373,6 +1373,8 @@ public partial class MainForm : Form
         }
     }
 
+    public void JoinJoyconToSelf(Joycon controller) => JoinJoycon(controller, controller);
+
     public void JoinJoycon(Joycon controller, Joycon other)
     {
         if (InvokeRequired)
